Proposal for refactor of pyutils
