+++ .gitignore
@@ -130,3 +130,5 @@ dmypy.json
 
 # Pycharm project
 .idea/
+
+venv/
