--- ebi_eva_common_pyutils/mongodb/__init__.py
@@ -1,1 +1,0 @@
-from ebi_eva_common_pyutils.mongodb.mongo_database import MongoDatabase
